requirement.txt
streamlit
pillow
pytesseract 
tesseract-ocr
tesseract-ocr-spa
